Refactor SocialMediaPosting class to extend BaseType and implement traits for enhanced functionality
